+++ core/execution/common/liquidity_provision.go
@@ -56,7 +56,7 @@ type MarketLiquidity struct {
 	marketID   string
 	asset      string
 
-	priceFactor *num.Uint
+	priceFactor num.Decimal
 
 	priceRange                num.Decimal
 	earlyExitPenalty          num.Decimal
@@ -77,7 +77,7 @@ func NewMarketLiquidity(
 	marketType marketType,
 	marketID string,
 	asset string,
-	priceFactor *num.Uint,
+	priceFactor num.Decimal,
 	priceRange num.Decimal,
 ) *MarketLiquidity {
 	ml := &MarketLiquidity{
@@ -788,42 +788,42 @@ func (m *MarketLiquidity) ValidOrdersPriceRange() (*num.Uint, *num.Uint, error) 
 	// (1 + priceRange) * midPrice
 	upperBoundPriceD := num.DecimalOne().Add(m.priceRange).Mul(midPrice)
 
-	priceFactor := m.priceFactor.ToDecimal()
-
 	// ceil lower bound
-	ceiledLowerBound, rL := lowerBoundPriceD.QuoRem(priceFactor, int32(0))
+	ceiledLowerBound, rL := lowerBoundPriceD.QuoRem(m.priceFactor, int32(0))
 	if !rL.IsZero() {
 		ceiledLowerBound = ceiledLowerBound.Add(num.DecimalOne())
 	}
-	lowerBoundPriceD = ceiledLowerBound.Mul(priceFactor)
+	lowerBoundPriceD = ceiledLowerBound.Mul(m.priceFactor)
 
 	// floor upper bound
-	flooredUpperBound, _ := upperBoundPriceD.QuoRem(priceFactor, int32(0))
-	upperBoundPriceD = flooredUpperBound.Mul(priceFactor)
+	flooredUpperBound, _ := upperBoundPriceD.QuoRem(m.priceFactor, int32(0))
+	upperBoundPriceD = flooredUpperBound.Mul(m.priceFactor)
 
 	lowerBound, _ := num.UintFromDecimal(lowerBoundPriceD)
 	upperBound, _ := num.UintFromDecimal(upperBoundPriceD)
 
 	// floor at 1 to avoid non-positive value
+	uintPriceFactor, _ := num.UintFromDecimal(m.priceFactor)
+
 	if lowerBound.IsNegative() || lowerBound.IsZero() {
-		lowerBound = m.priceFactor
+		lowerBound = uintPriceFactor
 	}
 
 	if lowerBound.GTE(upperBound) {
 		// if we ended up with overlapping upper and lower bound we set the upper bound to lower bound plus one tick.
-		upperBound = upperBound.Add(lowerBound, m.priceFactor)
+		upperBound = upperBound.Add(lowerBound, uintPriceFactor)
 	}
 
 	// we can't have lower bound >= best static ask as then a buy order with that price would trade on entry
 	// so place it one tick to the left
 	if lowerBound.GTE(bestAsk) {
-		lowerBound = num.UintZero().Sub(bestAsk, m.priceFactor)
+		lowerBound = num.UintZero().Sub(bestAsk, uintPriceFactor)
 	}
 
 	// we can't have upper bound <= best static bid as then a sell order with that price would trade on entry
 	// so place it one tick to the right
 	if upperBound.LTE(bestAsk) {
-		upperBound = num.UintZero().Add(bestAsk, m.priceFactor)
+		upperBound = num.UintZero().Add(bestAsk, uintPriceFactor)
 	}
 
 	return lowerBound, upperBound, nil
@@ -906,7 +906,7 @@ func NewMarketLiquidityFromSnapshot(
 	marketType marketType,
 	marketID string,
 	asset string,
-	priceFactor *num.Uint,
+	priceFactor num.Decimal,
 	state *snapshot.MarketLiquidity,
 ) *MarketLiquidity {
 	priceRange, _ := num.DecimalFromString(state.PriceRange)